Reorganized code in libgreenfield in anticipation of performance testing
